+++ server/src/controllers/food/delete-food.controller.ts
@@ -1,9 +1,14 @@
 import { Request, Response } from "express";
+import mongoose from "mongoose";
 import { FoodModel } from "../../models";
 
 export const deleteFood = async (req: Request, res: Response) => {
   try {
     const { foodId } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(foodId)) {
+      return res.status(400).json({ message: "ID буруу байна" });
+    }
 
     const deleted = await FoodModel.findByIdAndDelete(foodId);
 
@@ -12,6 +17,6 @@ export const deleteFood = async (req: Request, res: Response) => {
     return res.status(200).send({ message: "Food deleted", data: deleted });
   } catch (error) {
     console.error(error);
-    return res.status(500).send({ message: "Error deleting food", error });
+    return res.status(500).json({ message: "Серверийн алдаа гарлаа" });
   }
 };
